now speed can be a floating point value

+++ botSettings.js
@@ -2,7 +2,7 @@
 
 module.exports = {
     defaultOrbs: parseInt(process.env.DEFAULT_ORBS) || 10,
-    defaultSpeed: parseInt(process.env.BOT_SPEED) || 2,
+    defaultSpeed: parseFloat(process.env.BOT_SPEED) || 2,
     defaultSize: parseInt(process.env.BOT_SIZE) || 6,
     defaultZoom: parseInt(process.env.DEFAULT_ZOOM) || 2,
     worldWidth: parseInt(process.env.WORLD_WIDTH) || 1000,
